requirements.txt
Django==4.2.7
Pillow==10.1.0
django-environ==0.11.2
djangorestframework==3.14.0
django-cors-headers==4.3.1
celery==5.3.4
redis==5.0.1
reportlab==4.0.7
django-storages==1.14.2
boto3==1.34.0
stripe==7.8.0
razorpay==1.4.1
django-crispy-forms==2.1
crispy-bootstrap5==0.7
django-extensions==3.2.3
python-decouple==3.8